Fix theorem #4 implementation
I think this is right?

The product is still way larger than the actual result for values I tried.

We could probably determine if we got enough bits as we do the CRT stuff and
stop dynamically? That'd not be super helpful right now tho as we're only
parallel over different primes.